ci(Update): Change GetNormalizeByteSize return type to double
This change updates the return type of GetNormalizeByteSize from long to
double to handle fractional values during size normalization. The
absolute value calculation within the method was also modified to ensure
the conversion from long to double.

+++ MyExpenses.Models/IO/SizeDatabase.cs
@@ -70,9 +70,9 @@ public class SizeDatabase
         }
     }
 
-    private static long GetNormalizeByteSize(long bytes, out string unit)
+    private static double GetNormalizeByteSize(long bytes, out string unit)
     {
-        var absoluteBytes = Math.Abs(bytes);
+        var absoluteBytes = Math.Abs((double)bytes);
 
         string[] units = [
             SizeDatabaseResources.ByteUnit, SizeDatabaseResources.KiloByteUnit, SizeDatabaseResources.MegaByteUnit,
